ccai/tests/requirements.txt
pytest
python-googlegeocoder
flake8
